+++ backend/users/Touir.ExpensesManager.Users/Controllers/Requests/ResendVerificationRequest.cs
@@ -1,0 +1,8 @@
+namespace Touir.ExpensesManager.Users.Controllers.Requests
+{
+    public class ResendVerificationRequest
+    {
+        public required string Email { get; set; }
+        public required string ApplicationCode { get; set; }
+    }
+}