+++ backend/users/Touir.ExpensesManager.Users/Models/User.cs
@@ -11,6 +11,7 @@
         public DateTime LastUpdatedAt { get; set; }
         public bool IsEmailValidated { get; set; }
         public string? EmailValidationHash { get; set; }
+        public DateTime? EmailValidationHashExpiresAt { get; set; }
         public bool IsDisabled { get; set; }
         public bool IsDeleted { get; set; }
         public DateTime? DeletedAt { get; set; }